feat(utl): add absent entities
Defining an own type to denote absent elements can be avoided
passing the 'NoElem' empty type constructor,
defined in the 'Bem.Utl.Utl' module,
in place of a type constructor of them.
Defining an own type to denote absent modifiers can be avoided
passing the 'NoMod' empty type,
defined in the 'Bem.Utl.Utl' module,
in place of a type of them.